+++ test/CodeProject.ObjectPool.Benchmarks/Program.cs
@@ -37,7 +37,7 @@ namespace CodeProject.ObjectPool.Benchmarks
         {
             public Config()
             {
-                Add(Job.LegacyJitX86);
+                Add(Job.RyuJitX64);
                 Add(CsvExporter.Default, HtmlExporter.Default, MarkdownExporter.GitHub, PlainExporter.Default, CsvMeasurementsExporter.Default, RPlotExporter.Default);
                 Add(new MemoryDiagnoser());
                 Add(EnvironmentAnalyser.Default);