+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for CodeProject.ObjectPool #
 
+### v2.2.5 (2017-03-??) ###
+
+* Default minimum capacity is now 1.
+* Default maximum capacity is now 10.
+
 ### v2.2.4 (2017-03-05) ###
 
 * Fixed a bug which could produce closed pooled memory streams.